Add deep links, move files from shared
