+++ core-ui/src/components/Predefined/Details/StorageClasses/PersistentVolumeClaimsList.js
@@ -1,9 +1,20 @@
 import React from 'react';
+import LuigiClient from '@luigi-project/client';
+
 import { ComponentForList } from 'shared/getComponents';
 
 export function PersistentVolumeClaimsList(storageclass) {
+  const navigateFn = resource => {
+    LuigiClient.linkManager()
+      .fromContext('cluster')
+      .navigate(
+        `namespaces/${resource.metadata.namespace}/persistentvolumeclaims/details/${resource.metadata.name}`,
+      );
+  };
+
   const params = {
-    fixedPath: true,
+    hasDetailsView: true,
+    navigateFn,
     resourceUrl: `/api/v1/persistentvolumeclaims`,
     resourceType: 'PersistentVolumeClaims',
     isCompact: true,
