+++ shared/components/ResourcesList/ResourcesList.js
@@ -261,7 +261,7 @@ function Resources({
       <Link
         className="fd-link"
         onClick={_ => {
-          if (navigateFn) return navigateFn(entry.metadata.name);
+          if (navigateFn) return navigateFn(entry);
           if (fixedPath)
             return navigateToFixedPathResourceDetails(
               resourceType,
